+++ HISTORY.rst
@@ -7,6 +7,8 @@ History
 - Add processing algorithm to add NWRW parameters to surface parameters (#333)
 - Fix removing of cross-section locations associated to channels not modified by object import (#378)
 - Processing algorithms 'Map surfaces/DWF to connection nodes' now use point on surface instead of centroid as start vertex of mapping line (#396)
+- Fix handling string keys in import with value_map (#279)
+
 
 
 2.2.2 (2025-05-20)