+++ HISTORY.rst
@@ -10,6 +10,7 @@ History
 - Reset unspecified import settings on loading new import configuration (#286)
 - Fix handling string keys in import with value_map (#279)
 - Changing Material back to what it was does not update the friction value (#345)
+- Allow digitising and importing pipes with > 2 vertices (#342)
 
 
 2.2.2 (2025-05-20)